docs: Remove repo specific CONTRIBUTING.md
We now have a org wide CONTRIBUTING.md that points to our correct
general contributing guidelines.  We don't need repo specific ones that
forward to other contributing docs.

This is a fork but its CONTRIBUTING.md is very out of date and our fork
has veered of of upstream quite a bit so it doesn't make sense to keep
this file.